packages/ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy==2.0.0
Pillow==10.4.0
python-multipart==0.0.9
pydantic==2.9.0
